Rename previous insertData file
